Add tests
This doesn't feel right, as we're simply testing the Settings.Global.getFloat() method.

However, the bug it caused was tricky to track down and I suspect that some
userlying implemention has changed which caused something else other than
a Float to be returned, and that caused the entire app to crash.

So, putting in these tests should alert us to any further changes in future.